Extract components into files
